[YUNIKORN-591] move TaskGroupInstanceCountMap to its own file
