+++ pkg/common/utils/gang_utils.go
@@ -23,7 +23,6 @@ import (
 	"fmt"
 	"strconv"
 	"strings"
-	"sync"
 
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
@@ -144,74 +143,4 @@ func GetPlaceholderTimeoutParam(pod *v1.Pod) (int64, error) {
 		}
 	}
 	return 0, fmt.Errorf("no timeout parameter found")
-}
-
-type TaskGroupInstanceCountMap struct {
-	counts map[string]int32
-	sync.RWMutex
-}
-
-func NewTaskGroupInstanceCountMap() *TaskGroupInstanceCountMap {
-	return &TaskGroupInstanceCountMap{
-		counts: make(map[string]int32),
-	}
-}
-
-func (t *TaskGroupInstanceCountMap) Add(taskGroupName string, num int32) {
-	t.update(taskGroupName, num)
-}
-
-func (t *TaskGroupInstanceCountMap) AddOne(taskGroupName string) {
-	t.update(taskGroupName, 1)
-}
-
-func (t *TaskGroupInstanceCountMap) DeleteOne(taskGroupName string) {
-	t.update(taskGroupName, -1)
-}
-
-func (t *TaskGroupInstanceCountMap) update(taskGroupName string, delta int32) {
-	t.Lock()
-	defer t.Unlock()
-	if v, ok := t.counts[taskGroupName]; ok {
-		t.counts[taskGroupName] = v + delta
-	} else {
-		t.counts[taskGroupName] = delta
-	}
-}
-
-func (t *TaskGroupInstanceCountMap) Size() int {
-	t.RLock()
-	defer t.RUnlock()
-	return len(t.counts)
-}
-
-func (t *TaskGroupInstanceCountMap) GetTaskGroupInstanceCount(groupName string) int32 {
-	t.RLock()
-	defer t.RUnlock()
-	return t.counts[groupName]
-}
-
-func (t *TaskGroupInstanceCountMap) Equals(target *TaskGroupInstanceCountMap) bool {
-	if t == nil {
-		return t == target
-	}
-
-	t.RLock()
-	defer t.RUnlock()
-
-	if target == nil {
-		return false
-	}
-
-	if t.Size() != target.Size() {
-		return false
-	}
-
-	for k, v := range t.counts {
-		if target.counts[k] != v {
-			return false
-		}
-	}
-
-	return true
 }
